feat: daemon and server commands/alerts (#6)
This sets the stage for daemon <-> server communication via a gRPC
stream. I've tested this out running on the beta hardware running in k3s
and communication from the daemon (running on the host) to the server
(running in k3s and routed through envoy managed by fuse) works
flawlessly.