feat: App development API (#13)
This PR implements the operator logic to handle the new App development
API as structured in the new [store
repo](https://github.com/home-cloud-io/store).

It implements a simple App developer interface that provides:
- Persistence for data using local PV/PVCs
- Databases (for only Postgres at the moment) with separate users,
databases, and initialization scripts
- Secrets that can be generated at application install time
- Routes that configure both the routing in Envoy through Fuse as well
as set up mDNS for local connections

There's a few things in here to make this work:
- A new service was added called `mdns` that watches for k8s `Services`
of type `ExternalName` in the `home-cloud` namespace and configures mDNS
based on the service names (with the shape
`serviceName-home-cloud.local`). The `mdns` service must have both the
`HOST_IP` and the `HOST_NAME` given to it in the form of env vars (we
should change these to configs at some point) and it uses these to only
advertise mDNS values that match it's own `HOST_IP`. This means we can
run it as a `DaemonSet` when we cluster more nodes.
- The `operator` had a bunch of functionality added to it like creating
all the needed k8s objects (Secrets, PV/PVCs, etc.) as well as a
database client that it uses to initialize databases and create users.
Currently a delete call is non-destructive of these objects and the
database values and we'll need to add a "hard delete" at some point.
- The server just had some minor cleanup, pointed the web-client at the
new app store and enabled HTTP2 on its Fuse registration.