feat: App dependencies (#44)
Added logic to the App installation on the server that checks
dependencies (as defined in the Chart) and installs them before
installing the requested app. The primary usecase for this is databases
like Postgres for Immich.

I also cleaned up the models for the App Store cache and put it on a
background refresh timer of 24hr.