feat: auto update (#46)
The core feature in this PR is automatic updates for the OS, the daemon,
system containers, and Apps.

Alongside this is a complete refactor of the server code base. I was
running into issues with cyclic imports because we had our wires crossed
with the way we put code into packages. This PR moves things around to
push feature sets (`apps`, `system`, etc.) into separate buckets to
avoid mixing things like it was before (e.g. `versioning`).

I also cleaned up a bunch of interfaces, built a key-value store client,
and consolidated the various runners into a single function within
`main`.