feat: daemon version hashes (#55)
Forgot when working on auto updates to go back and resolve a TODO item
about retrieving the daemon hashes (vendor and src) for the latest
version to install. These are needed to perform upgrades and I'd been
manually pushing them through the API to update the daemon up until now.

This PR adds logic to the server and daemon to process the version and
source code/vendor hashes that define a full daemon version. The latest
hashes are obtained by the server by evaluating git tags in the form of
`daemon_VERSION_src_HASH` and `daemon_VERSION_vendor_HASH`. This is a
bit clunky but gets us automatic updates for the daemon for now which is
essential for RC1.