feat: better mDNS (#72)
This one will be a story...

We need an implementation of mDNS for two reasons. We need the Home
Cloud host itself to be visible on the LAN as `home-cloud.local` so that
the user can access the dashboard and we need each App to be able to
register hosts on the LAN as well (like `photos.local` for Immich).

I put in a ton of work and research on this and came up with a solution
that *should* have worked using the
[pion/mdns](https://github.com/pion/mdns) module but it was flaky. For
the most part it worked for iOS, OSX, and Windows devices but it would
almost never work for Android. I debugged the thing all the way down the
stack and as far as I could tell everything was functioning but every
Android device I tested across two different networks wouldn't work
(except for some rare exceptions which I still don't understand).

Eventually I decided to test just using Avahi instead and it's been rock
solid for all my devices. It means managing mDNS through the `daemon`
instead of the `mdns` microservice which removes a component from the
system which I guess is good for simplicity. It also means we don't have
a pod in k3s that's sitting on the host network which reduces our
security surface.

This PR fully implements mDNS using the new structure of calling the
`avahi-publish` command by the daemon and supports both adding and
removing hostnames when apps are added and removed from the system.

**NOTE**: While this PR fully supports removing routes from `fuse`, that
functionality is not yet implemented in `fuse` itself.