fix: release child command on `nixos-rebuild switch` (#84)
Fixes a bug where the `daemon` is killed by systemd while executing the
`nixos-rebuild switch` command and this also kills the child command so
services stopped during the application of the OS update are not
restarted. This causes the `daemon`, networking, k3s (any service
needing an update really) to stay in a down state requiring a reboot of
the machine to function again.

This PR fixes this issue by releasing the underlying process running the
`nixos-rebuild switch` command so that when the `daemon` is stopped the
underlying command will keep running. The drawback is that we lose the
output/status of the underlying command so we have no way of knowing
about the command's success or failure. We can revisit this issue later.