fix: uint64 stats (#114)
In a previous PR this was changed to uint32 because of some confusion
around precision between TS and Golang. While uint32 is enough for TS,
Golang needs uint64. Multi-language support is fun. Anyway, changing it
back and will do a conversion in TS since the proto gets compiled to a
`bigint` in TS.