feat: JSON-based Nix configuration (#115)
Added a migration to the daemon to shift the NixOS configuration to
`.json` based files instead of `.nix` so that the daemon can manage the
full configuration without resorting to the string parsing replacement
methods. Also fixed up some bugs left over from the previous PR. Still
need to go back and do the same type of change for k3s manifests.