fix: improvements to file upload (#116)
While there's still some networking/envoy issue causing uploads to fail
sporadically, this PR includes a bunch of improvements to the upload
process. There's a rework to the listener/broadcaster pattern that makes
it non-blocking, more efficient, and provides a buffering pattern for
slower event listeners. The daemon now handles file uploads in a
stateless manner that is also non-blocking.