fix: daemon version bugs (#120)
Fixes a bug where the daemon nix file path is incorrect causing errors
when reading and updating the version. Also fixed the server to properly
respond to the error returned by the daemon when this occurred. And
finally, fixed a bad comparison which was pointer compare instead of
attribute compare.