Write peer wireguard configuration to the interface (#127)
When the daemon receives a message to add a wireguard peer it's added to
the nix configuration and then `Rebuilt` and `Switched`.

NOTE: The client is unaware of the wireguard interface to add the new
peer to so by default it's being added to each `wg` interface. That will
change when the device peering feature is created.