feat: Data Drive Support (#128)
Now that we're moving to mount a data drive to separate the OS from user
space data, this updates the daemon to use that drive's mount point
instead of root. Also cleaned up the storage text in the UI and added an
experimental badge to the On the Go settings.

NOTE: Also changed the daemon to be on go `1.22.7` since that's the
latest it seems that NixOS supports and the devices have to install it
from source.