feat: Wireguard tunneling proto changes (#149)
Moving around some of the protos dealing with Wireguard tunneling. This
simplifies things and restructures the API between the server and the
daemon so that the proper hierarchy between Wireguard interface,
locators, and STUN servers is used.

Basically, a Wireguard interface has its own configured STUN server and
its own list of locators whereas before the protos were defined assuming
a global list of locators, a single global STUN server, and Wireguard
interfaces attached to locators (the reverse of what we really need).