feat: end-to-end Wireguard (#146)
This PR implements an end-to-end Wireguard tunnel with NAT hole punching
negotiated using a Locator server. This PR includes the necessary
changes to the daemon, server, and locator client example to demonstrate
the technique end to end.

Process:

*setup (on local network)*
- Client tells the server to enable secure tunneling and the server
communicates with the daemon which configures the NixOS to set up the
Wireguard server on the host.
- The client then can tell the server to register with one or many
Locator servers and again the server communicates with the daemon to
configure these connections.
- The client can then request a Wireguard peer to be added and after the
daemon configures the host all the necessary connection information is
returned to the client.


*connection (on remote network)*
- The client sends a request to one of the configured Locator servers to
retrieve the STUN information of the server.
- The daemon verifies the request using the previously exchanged
Wireguard encryption keys and returns its STUN information.
- The client and daemon both start sending packets at each other to open
a hole in the NAT to each others' STUN addresses.
- The client simultaneously starts its Wireguard server with its peer
configuration for the server pointing to the STUN address retrieved
earlier.
- The Wireguard tunnel is established and is kept alive via Wireguard
pings every 25 seconds.

I've tested and verified this entire process from a mobile network to my
home network with a Locator server running in GCP handling both Locator
and STUN processing.

Some other changes worth noting:
- Updates to Go 1.23.6 for both daemon and server
- Merging the Wireguard settings into the main server settings as a
child attribute
- Consolidation of STUN, Locator, and Wireguard methods in the daemon
into controllers
- Simplification of Daemon->Server error handling in protos
- Removal of writing Wireguard private keys to Blueprint
- Implementation of ability to host multiple independent Wireguard
servers on host through Daemon