feat: Tunneling dashboard settings (#153)
There's a few changes in this PR. The first is the addition of multiple
drive support and system stats on the Home page. This is just a UI
change as the API already returned all this info before:


![image](https://github.com/user-attachments/assets/6e17b2d4-9c91-489f-bae2-8ee39b879c36)

The second change is a rework of the On the Go Settings card to work
with the new API for secure tunneling. This now looks a bit cleaner by
using a modal for adding locators and has support for both default and
custom Locator servers:


![image](https://github.com/user-attachments/assets/ccbf527f-4d73-4eaf-a121-25de2d5c1320)


![image](https://github.com/user-attachments/assets/8cdd8468-7245-48f9-b4c3-27f2b9ac1645)


![image](https://github.com/user-attachments/assets/6e43f24f-e702-414e-8cda-84e3d417e1da)

Finally, there is also a fundamental change to the way events are passed
between the server and the daemon. I added a new attribute to every
message called the `subject` in #152. This enables a request/response
pattern between the server and daemon which I wrapped in the
`Commander.Request()` method to clean up a lot of the clunky
`async.Listener` code that was being used throughout the server. When
calling `Commander.Request()`, a UUID is generated for the message
`subject` and sent to the daemon which then adds the same UUID as the
`subject` on any return messages.