fix: daemon Darwin build support (#155)
Since the daemon is only designed to be run on Linux systems, the OS
stat collection was using Linux-only calls from the
`github.com/mackerelio/go-osstat` module. This PR splits the stats
collection into separate `linux` and `darwin` files so that the daemon
can be built (for development only) on Mac systems.